refactor: 불필요한 파일 및 폴더 삭제
